Move BTC_TX_VERSION constants to BitcoinUtils
