+++ rskj-core/src/main/java/co/rsk/peg/ReleaseTransactionBuilder.java
@@ -18,38 +18,42 @@
 
 package co.rsk.peg;
 
-import co.rsk.bitcoinj.core.*;
+import static co.rsk.peg.PegUtils.getFlyoverFederationAddress;
+import static co.rsk.peg.bitcoin.BitcoinUtils.addSpendingFederationBaseScript;
+import static com.google.common.base.Preconditions.checkNotNull;
+
+import co.rsk.bitcoinj.core.Address;
+import co.rsk.bitcoinj.core.BtcTransaction;
+import co.rsk.bitcoinj.core.Coin;
+import co.rsk.bitcoinj.core.InsufficientMoneyException;
+import co.rsk.bitcoinj.core.NetworkParameters;
+import co.rsk.bitcoinj.core.TransactionInput;
+import co.rsk.bitcoinj.core.UTXO;
+import co.rsk.bitcoinj.core.UTXOProviderException;
 import co.rsk.bitcoinj.script.Script;
 import co.rsk.bitcoinj.wallet.RedeemData;
 import co.rsk.bitcoinj.wallet.SendRequest;
 import co.rsk.bitcoinj.wallet.Wallet;
 import co.rsk.crypto.Keccak256;
+import co.rsk.peg.bitcoin.BitcoinUtils;
 import co.rsk.peg.federation.Federation;
 import co.rsk.peg.federation.FederationFormatVersion;
+import java.util.List;
 import org.ethereum.config.blockchain.upgrades.ActivationConfig;
 import org.ethereum.config.blockchain.upgrades.ConsensusRule;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.*;
-
-import static co.rsk.peg.PegUtils.getFlyoverFederationAddress;
-import static co.rsk.peg.bitcoin.BitcoinUtils.addSpendingFederationBaseScript;
-import static com.google.common.base.Preconditions.checkNotNull;
-
 /**
  * Given a set of UTXOs, a ReleaseTransactionBuilder
  * knows how to build a release transaction
- * of a certain amount to a certain address,
+ * of a certain amount to a certain address
  * and how to signal the used UTXOs so they
  * can be invalidated.
  *
  * @author Ariel Mendelzon
  */
 public class ReleaseTransactionBuilder {
-
-    public static final int BTC_TX_VERSION_1 = 1;
-    public static final int BTC_TX_VERSION_2 = 2;
 
     public record BuildResult(BtcTransaction btcTx, List<UTXO> selectedUTXOs, Response responseCode) {}
 
@@ -132,7 +136,7 @@ public class ReleaseTransactionBuilder {
     public BuildResult buildMigrationTransaction(Coin migrationValue, Address destinationAddress) {
         return buildWithConfiguration((SendRequest sr) -> {
             if (!activations.isActive(ConsensusRule.RSKIP376)){
-                sr.tx.setVersion(BTC_TX_VERSION_1);
+                sr.tx.setVersion(BitcoinUtils.BTC_TX_VERSION_1);
             }
             sr.tx.addOutput(migrationValue, destinationAddress);
             sr.changeAddress = destinationAddress;
@@ -196,7 +200,7 @@ public class ReleaseTransactionBuilder {
         // Build a tx and send request and configure it
         BtcTransaction btcTx = new BtcTransaction(params);
         if (activations.isActive(ConsensusRule.RSKIP201)) {
-            btcTx.setVersion(BTC_TX_VERSION_2);
+            btcTx.setVersion(BitcoinUtils.BTC_TX_VERSION_2);
         }
 
         return btcTx;
